+++ api/controllers/controllers.transaction.js
@@ -1,27 +1,5 @@
 const transactionService = require('../services/services.transaction')
 
-const checkBalanceController = async (req, res) => {
-  const { clientid } = req.params
-  const resultService = await transactionService.checkBalanceService(clientid)
-  const code = resultService.success ? 200 : 400
-  const message = resultService.success
-    ? { message: resultService.message }
-    : { details: resultService.details }
-  const data = resultService.data ? resultService.data : ''
-  return res.status(code).send({ message, data })
-}
-
-const listUserAssetController = async (req, res) => {
-  const { clientid } = req.params
-  const resultService = await transactionService.listUserAssetService(clientid)
-  const code = resultService.success ? 200 : 400
-  const message = resultService.success
-    ? { message: resultService.message }
-    : { details: resultService.details }
-  const data = resultService.data ? resultService.data : ''
-  return res.status(code).send({ message, data })
-}
-
 const createTransactionController = async (req, res) => {
   const { body } = req
   const { params } = req
@@ -39,7 +17,5 @@ const createTransactionController = async (req, res) => {
 }
 
 module.exports = {
-  checkBalanceController,
-  listUserAssetController,
   createTransactionController
 }
