+++ api/routers/v1/routers.account.js
@@ -1,0 +1,17 @@
+const joi = require('joi')
+const accountController = require('../../controllers/controllers.account')
+const middlewareValidateDTO = require('../../utils/middlewares/middlewares.validate_dto')
+const authorizationMiddleware = require('../../utils/middlewares/middlewares.authorization')
+
+module.exports = (router) => {
+  router.route('/client/:clientid/account').get(
+    authorizationMiddleware('LIST_CLIENT_BALANCE'),
+    middlewareValidateDTO('params', {
+      clientid: joi.number().integer().required().messages({
+        'any.required': '"client id" is a required field',
+        'string.empty': '"client id" can not be empty'
+      })
+    }),
+    accountController.checkBalanceController
+  )
+}
