+++ api/routers/v1/routers.client.js
@@ -11,6 +11,18 @@ module.exports = (router) => {
       authorizationMiddleware('SEARCH_FINANCIAL'),
       clientController.listAllClientsController
     )
+
+  router.route('/client/:clientid/assets').get(
+    authorizationMiddleware('LIST_CLIENT_ASSETS'),
+    middlewareValidateDTO('params', {
+      clientid: joi.number().integer().required().messages({
+        'any.required': '"client id" is a required field',
+        'string.empty': '"client id" can not be empty'
+      })
+    }),
+    clientController.listUserAssetController
+  )
+
   router
     .route('/client/:clientid')
     .get(
